+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jala-banyu/input
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.1.0
+
 ## 1.0.0
 
 ### Major Changes