+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/popover
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.1.0
+  - @jala-banyu/button@2.0.0
+
 ## 1.0.0
 
 ### Major Changes